Start of downloader and functional test
Starting to put together the plugin. Unfortunately, I think the
currently logic is pretty broken because it needs to handle deferred
setting of values through gradle.

Downloader should take providers of strings and return a project.file of
the target directory, maybe? That then needs to be added to the test
configuration, and test needs to depend on downloading jpf.